fix(minor): lockdown frappe.client.get_list
