+++ frappe/desk/reportview.py
@@ -41,6 +41,9 @@ def get_form_params():
 	"""Stringify GET request parameters."""
 	data = frappe._dict(frappe.local.form_dict)
 	clean_params(data)
+	validate_args(data)
+
+def validate_args(data):
 	parse_json(data)
 	setup_group_by(data)
