+++ frappe/model/db_query.py
@@ -14,7 +14,6 @@ import frappe.permissions
 from datetime import datetime
 import frappe, json, copy, re
 from frappe.model import optional_fields
-from frappe.client import check_parent_permission
 from frappe.model.utils.user_settings import get_user_settings, update_user_settings
 from frappe.utils import flt, cint, get_time, make_filter_tuple, get_filter, add_to_date, cstr, get_timespan_date_range
 from frappe.model.meta import get_table_columns
@@ -785,6 +784,18 @@ class DatabaseQuery(object):
 			user_settings['fields'] = self.user_settings_fields
 
 		update_user_settings(self.doctype, user_settings)
+
+def check_parent_permission(parent, child_doctype):
+	if parent:
+		# User may pass fake parent and get the information from the child table
+		if child_doctype and not frappe.db.exists('DocField',
+			{'parent': parent, 'options': child_doctype}):
+			raise frappe.PermissionError
+
+		if frappe.permissions.has_permission(parent):
+			return
+	# Either parent not passed or the user doesn't have permission on parent doctype of child table!
+	raise frappe.PermissionError
 
 def get_order_by(doctype, meta):
 	order_by = ""
